refactor: 도메인 함수 리팩토링 및 디렉토리 구조 수정
