+++ src/js/domain/Search.js
@@ -1,5 +1,6 @@
-import storage from './storage';
 import { SEARCH_URL_BASE, MAX_SEARCH_RESULT, ERROR_MESSAGES } from '../constants/constants';
+import getVideoObjectArray from './utils/getVideoObjectArray';
+import { getFromStorage } from './storage';
 
 class Search {
   #keyword;
@@ -16,11 +17,10 @@ class Search {
       const { items, nextPageToken } = await this.#getSearchResult(keyword, this.#nextPageToken);
       this.#keyword = keyword;
       this.#nextPageToken = nextPageToken;
-      const savedVideos = [
-        ...storage.getFromStorage('watched'),
-        ...storage.getFromStorage('unwatched'),
-      ];
-      const searchResultArray = this.#getVideoObjectArray(items, savedVideos);
+
+      const savedVideos = [...getFromStorage('watched'), ...getFromStorage('unwatched')];
+      const searchResultArray = getVideoObjectArray(items, savedVideos);
+
       return { searchResultArray, hasNextPage: !!nextPageToken };
     } catch (error) {
       throw new Error(error.message);
@@ -29,10 +29,13 @@ class Search {
 
   async #getSearchResult(keyword, pageToken) {
     const queryString = this.#generateQueryString(keyword, pageToken);
+
     try {
       const response = await fetch(`${SEARCH_URL_BASE}${queryString}`);
-      if (!response.ok) throw new Error();
+      if (!response.ok) throw new Error(response);
+
       const { items, nextPageToken } = await response.json();
+
       return { items, nextPageToken };
     } catch (error) {
       throw new Error(ERROR_MESSAGES.SERVER_ERROR);
@@ -40,30 +43,17 @@ class Search {
   }
 
   #generateQueryString(keyword, pageToken) {
-    const query = {
+    const queryString = new URLSearchParams({
       part: 'snippet',
       type: 'video',
       regionCode: 'KR',
-      q: keyword,
       maxResults: MAX_SEARCH_RESULT,
-    };
-    if (pageToken) query.pageToken = pageToken;
-    return new URLSearchParams(query);
-  }
-
-  #getVideoObjectArray(items, savedVideos) {
-    if (items.length === 0) throw new Error(ERROR_MESSAGES.NO_RESULT);
-    return items.map((item) => {
-      const { snippet, id } = item;
-      return {
-        videoId: id.videoId,
-        thumbnail: snippet.thumbnails.medium.url,
-        title: snippet.title,
-        channelTitle: snippet.channelTitle,
-        publishedAt: snippet.publishedAt,
-        isSaved: !!savedVideos.includes(id.videoId),
-      };
+      q: keyword,
     });
+
+    if (pageToken) queryString.append('pageToken', pageToken);
+
+    return queryString;
   }
 }
